Reorder eslint plugins in config
The order of eslint plugins in the .eslintrc.js file has been changed, moving 'plugin:import/recommended' and 'plugin:import/typescript' lower in the list of plugins. This change has been made to attempt to resolve a conflict between these plugins and other plugins that were previously lower in the execution order. The resolution of this conflict should lead to more consistent linting results and fewer false positives.