Update ESLint rule and bump up version
Changed the rule for "functional/immutable-data" in `.eslintrc.js` to specifically ignore class fields. This change is made to improve the flexibility of the rule and to support certain use cases that involve modifying class fields.

Simultaneously, updated the `version` in the `package.json` file from `10.1.0` to `10.1.1`. This is done to indicate the introduction of these new changes in the ESLint rule and to maintain accurate version control.