Update ESLint rules and remove '@total-typescript/ts-reset'
New rules have been added to .eslintrc.js to enforce more type checking in the code. The '@total-typescript/ts-reset' package was removed from index.js, package.json, and pnpm-lock.yaml as it was no longer needed. Overall, these changes aim to improve the quality and robustness of the code base.